+++ nacc_attribute_deriver/attributes/missingness/modules/uds/missingness_a2.py
@@ -5,6 +5,10 @@ from nacc_attribute_deriver.utils.constants import INFORMED_MISSINGNESS
 
 
 class UDSFormA2Missingness(UDSMissingness):
+    def _missingness_langa2(self) -> int:
+        """Handle missingness for LANGA2."""
+        return self.handle_optional_header_variables("lang", "a2")
+
     ############################
     # INLIVWTH-gated variables #
     ############################