+++ tests/attributes/derived/modules/uds/test_form_header.py
@@ -80,9 +80,6 @@ class TestUDSHeaderAttributeCollection:
     def test_create_naccdays(self, table, form_prefix, working_derived_prefix):
         """Tests _create_naccdays."""
         attr = UDSHeaderAttributeCollection(table)
-
-        # actually 16,462 but maxes out at 5000
-        # assert attr._create_naccdays() == 5000
         assert attr._create_naccdays() == 16462
 
         # set it closer
@@ -104,7 +101,14 @@ class TestUDSHeaderAttributeCollection:
         set_attribute(table, form_prefix, "packet", "IT")
         assert attr._create_naccdays() == 0
 
+    def test_create_naccdays_i4(self, table, form_prefix):
+        """Tests _create_naccdays on an I4 visit."""
+        set_attribute(table, form_prefix, "packet", "I4")
+        attr = UDSHeaderAttributeCollection(table)
+
+        assert attr._create_naccdays() == 16462
+
     def test_create_naccnvst(self, table):
-        """Tsests _create_naccnvst."""
+        """Tests _create_naccnvst."""
         attr = UDSHeaderAttributeCollection(table)
         assert attr._create_naccnvst() == 5